Add delete confirmation dialog
Former-commit-id: 98dac34bcbe357daba42025130eacfd2575a834d

+++ public/javascripts/ember/controllers.js
@@ -1,9 +1,14 @@
 App.CarsController = Ember.ObjectController.extend({
 	actions: {
 		delete: function(carId) {
-			this.store.find('car', carId).then(function(car) {
-				car.deleteRecord();
-				car.save();	
+			var that = this;
+			bootbox.confirm('האם למחוק?', function(userWantsToDelete) {
+				if (userWantsToDelete) {
+					that.store.find('car', carId).then(function(car) {
+						car.deleteRecord();
+						car.save();	
+					});
+				}
 			});
 		}
 	}
